+++ src/main/java/com/consol/api/repository/InstituicaoRepository.java
@@ -1,5 +1,6 @@
-package com.example.APIcrudconsol.instituicao;
+package com.consol.api.repository;
 
+import com.consol.api.entity.Instituicao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstituicaoRepository extends JpaRepository<Instituicao, Integer> {
