+++ src/main/java/com/example/APIcrudconsol/necessidade/Necessidade.java
@@ -1,6 +1,6 @@
 package com.example.APIcrudconsol.necessidade;
 
-import com.example.APIcrudconsol.instituicao.Instituicao;
+import com.consol.api.entity.Instituicao;
 import jakarta.persistence.*;
 import lombok.*;
